feat(ui): split EntityBreadcrumbPureUtils into domain-specific files
Extract 1046-line monolith EntityBreadcrumbPureUtils.ts into focused
domain files:
- EntityLinkUtils.ts: getEntityLinkFromType
- EntityDataBreadcrumbUtils.ts: data asset breadcrumbs (table, chart, API, container)
- EntityGovernanceBreadcrumbUtils.ts: governance breadcrumbs (glossary, domain, KPI, test, etc.)
- EntityServiceBreadcrumbUtils.ts: service breadcrumbs (database, schema, service category)

EntityBreadcrumbPureUtils.ts becomes a re-export facade keeping
getEntityBreadcrumbs dispatcher. EntityBreadcrumbUtils.tsx re-exports
all symbols for full backward compatibility.
